Workout History and Custom Exercise Implemented
The strength log was changed first cardio log has not been changed yet. A list of exercises to choose from was implemented using a master list of exercises in a dropdown menu to make it easier to graph exercises in the future. A user can add their own custom exercise to this list. A user can click on a date on the calendar and see the strength and cardio logs for that date as a workout history.